move Tee, processProfileStats, makeProfileGraph to profile module
remove deprecated imports
